requirements.txt
netifaces
websocket-client
rel
pyserial
pigpio
pywiegandpi >= 1.0.1
